Refactor: extract DB connection into singleton db.ts pattern (fixes #49)
- Aligned exactly with architect's recommended db.ts singleton pattern
- Extracted queryLocalDb, getDbPath, and resetDbConnection into backend/db.ts
- Re-exported queryLocalDb from server.ts to preserve existing feature imports
